requirements.txt
crewai
langchain-community
langchain-openai
requests
duckduckgo-search
chromadb
google-generativeai
langchain-groq
streamlit